+++ psrc/t8dev/cli/t8dev/emulator/__init__.py
@@ -1,14 +1,14 @@
 ''' cli.t8dev emulator commands
 
-    e help              help function for emulation commands
-    e list              lists all of the suites available e.g. cscp, vice etc
-    e cscp list         lists all available emulators within suite
-    e cscp tk85         specific emulator from cscp suite
-    e cscp pc8001
-    e vice x64
-    e openmsx           suite should have multiple MSX machine in it?
-    e linapple
-    e runcpm            probably no suite here, because just the one
+    emu help            help function for emulation commands
+    emu list            lists all of the suites available e.g. cscp, vice etc
+    emu cscp list       lists all available emulators within suite
+    emu cscp tk85       specific emulator from cscp suite
+    emu cscp pc8001
+    emu vice x64
+    emu openmsx         suite should have multiple MSX machine in it?
+    emu linapple
+    emu runcpm          probably no suite here, because just the one
 '''
 
 from    pathlib  import Path
@@ -29,28 +29,32 @@ import  t8dev.run  as run
 
 ####################################################################
 
-def emulator(args):
-    '''
-    '''
-    if not args:
-        exits.arg("arguments required; 'help' for help")
-    if args[0] == 'help':
-        print('Usage: t8dev emulate SUITE [emulator] [args...]')
-        print('       t8dev emulate list    # list emulator suites')
-        exit(0)
-    if args[0] == 'list':
-        print('Emulator suites:')
-        for c in sorted(SUITES): print(f'  {c}')
-        exit(0)
-    cls = SUITES.get(args[0], None)
-    if cls is None:
-        exits.arg(f"Bad suite name '{args[0]}'. Use 'list' for list of suites.")
-    suite = cls(args[1:])
-    suite.run()
+def setargs_emulator(spgroup):
+    pemu = spgroup.add_parser('emulator', aliases=['emulate', 'emu', 'em'],
+        help='run an emulator')
+    suitegroup = pemu.add_subparsers(dest='suite', required=True,
+        title='Emulator suites', metavar='', help='')
+    for sc in SUITE_CLASSES:
+        p = suitegroup.add_parser(sc.suitename(), help=sc.suitedesc)
+        p.set_defaults(func=run_emulator(sc))
+        sc.setargparser(p)
 
-####################################################################
+def run_emulator(suiteclass):
+    '   XXX This seems a bit of a hack.... '
+    def run(args): e = suiteclass(args); return e.run()
+    return run
 
 class Suite:
+
+    @classmethod
+    def suitename(cls):
+        return cls.__name__.lower()
+
+    @classmethod
+    def setargparser(cls, parser):
+        ''' This is given a parser for the subcommand for this
+            emulator suite; override this to add add arguments to it.
+        '''
 
     def __init__(self, args):
         self.args = args
@@ -58,67 +62,23 @@ class Suite:
     def run(self):
         exits.arg(f'{self}: unimplemented (args={self.args})')
 
-    @classmethod
-    def suitename(cls):
-        return cls.__name__.lower()
-
 class CSCP(Suite):
-
     #   XXX This requires that wine be installed (on Linux) and the
     #   emulators be extracted to .build/tool/bin/cscp/, which is
     #   done by `t8dev buildtoolset cscp`. We should give a hint
     #   about doing that if the emulators are not present.
 
-    #   VENDOR_ROM entries may be for an actual CSCP emulator name, or may
-    #   be alternate configurations that load a different set of ROMs
-    #   (e.g., `tk80` below). The alternate ROM configurations include an
-    #   `_emulator` entry which is not a ROM image but indicates the actual
-    #   name of the emulator to use (e.g., `tk80bs` for the `tk80` entry
-    #   below).
-    VENDOR_ROM = {
-        'pc8001': {
-            'N80.ROM': 'https://gitlab.com/retroabandon/pc8001-re/-/raw/main/rom/80/N80_11.bin',
-            'KANJI1.ROM': '@1000:https://gitlab.com/retroabandon/pc8001-re/-/raw/main/rom/80/FONT80.bin',
-        },
-        'tk80': {
-            '_emulator': 'tk80bs',
-            'TK80.ROM':             # $0000-$07FF fill $FF
-                'https://gitlab.com/retroabandon/tk80-re/-/raw/main/rom/TK80.bin',
-            'EXT.ROM': None,        # $0C00-$7BFF
-            'BSMON.ROM': None,      # $F000-$FFFF
-        },
-        'tk80bs': {
-            #   With TK80.ROM not present (file size 0) the emulator sets
-            #   `jp $F000` at RST 0 and `jp $83DD` at RST 7, like the
-            #   `tk80.dummy` ROM that MAME uses.
-            'TK80.ROM': None,       # $0000-$07FF fill $FF
-            'EXT.ROM':              # $0C00-$7BFF
-                'https://gitlab.com/retroabandon/tk80-re/-/raw/main/rom/80BS-clone/etkroms3/Ext.rom',
-            #   If this ROM _is_ present (file size >0) the emulator sets
-            #   `jp $83DD` at RST 7 (patching TK80.ROM if present).
-            'BSMON.ROM':            # $F000-$FFFF
-                'https://gitlab.com/retroabandon/tk80-re/-/raw/main/rom/80BS-clone/etkroms3/bsmon.rom',
-            #   Only one of LV[12]BASIC.ROM is read based on the boot mode.
-            'LV1BASIC.ROM':         # $E000-$EFFF
-                'https://gitlab.com/retroabandon/tk80-re/-/raw/main/rom/80BS-clone/etkroms/L1CB.ROM',
-            'LV2BASIC.ROM':         # $D000-$EFFF
-                'https://gitlab.com/retroabandon/tk80-re/-/raw/main/rom/80BS-clone/etkroms3/LV2BASIC.ROM',
-            'FONT.ROM':             # chargen addr space
-                'https://gitlab.com/retroabandon/tk80-re/-/raw/main/rom/80BS-clone/etkroms3/font.rom',
-        },
-        'tk85': {
-            'TK85.ROM':             # $0000-$07FF fill $FF
-                'https://gitlab.com/retroabandon/tk80-re/-/raw/main/rom/TK85.bin',
-            'EXT.ROM': None,        # $0C00-$7BFF
-        },
-    }
+    suitedesc = 'Comon Source Code Project emulators'
+
+    @classmethod
+    def setargparser(cls, parser):
+        parser.add_argument('emulator',
+            help="the emulator to run (use 'list' to list all emulators)")
+        parser.add_argument('patchspecs', nargs='*', help='ROM patchspecs')
 
     def run(self):
-        if not self.args:
-            exits.arg("Missing emulator name. Use 'list' for list of emulators.")
-
         self.set_bindir()
-        self.emulator = emulator = self.args.pop(0)
+        self.emulator = emulator = self.args.emulator
         if emulator == 'list':
             self.print_emulators()
             return
@@ -175,14 +135,16 @@ class CSCP(Suite):
                 continue
             try:
                 ri = RomImage(filename, path.download('rom-image'), loadspec)
-                ri.patches(self.args)   # removes args it used and patched
+                #   Use applicable patchspecs and remove from list.
+                ri.patches(self.args.patchspecs) 
             except FileNotFoundError as ex:
                 exits.err(ex)
             except HTTPError as ex:
                 exits.err(f'{ex}: {ex.url}')
             ri.writefile(self.emudir(filename))
-        if self.args:
-            exits.arg('Unknown arguments:', *[ f'  {arg}' for arg in self.args ])
+        if self.args.patchspecs:
+            unknowns = ' '.join(self.args.patchspecs)
+            exits.arg(f'Unknown patchspecs for {self.emulator}: {unknowns}')
 
         self.emudir(emuexe).unlink(missing_ok=True)
         #   Wine emulates Windows *really* well and throws up on
@@ -198,32 +160,79 @@ class CSCP(Suite):
         toolset.setpath()               # should be doing this?
         self.bindir = toolset.bindir()
 
+    #   VENDOR_ROM entries may be for an actual CSCP emulator name, or may
+    #   be alternate configurations that load a different set of ROMs
+    #   (e.g., `tk80` below). The alternate ROM configurations include an
+    #   `_emulator` entry which is not a ROM image but indicates the actual
+    #   name of the emulator to use (e.g., `tk80bs` for the `tk80` entry
+    #   below).
+    VENDOR_ROM = {
+        'pc8001': {
+            'N80.ROM': 'https://gitlab.com/retroabandon/pc8001-re/-/raw/main/rom/80/N80_11.bin',
+            'KANJI1.ROM': '@1000:https://gitlab.com/retroabandon/pc8001-re/-/raw/main/rom/80/FONT80.bin',
+        },
+        'tk80': {
+            '_emulator': 'tk80bs',
+            'TK80.ROM':             # $0000-$07FF fill $FF
+                'https://gitlab.com/retroabandon/tk80-re/-/raw/main/rom/TK80.bin',
+            'EXT.ROM': None,        # $0C00-$7BFF
+            'BSMON.ROM': None,      # $F000-$FFFF
+        },
+        'tk80bs': {
+            #   With TK80.ROM not present (file size 0) the emulator sets
+            #   `jp $F000` at RST 0 and `jp $83DD` at RST 7, like the
+            #   `tk80.dummy` ROM that MAME uses.
+            'TK80.ROM': None,       # $0000-$07FF fill $FF
+            'EXT.ROM':              # $0C00-$7BFF
+                'https://gitlab.com/retroabandon/tk80-re/-/raw/main/rom/80BS-clone/etkroms3/Ext.rom',
+            #   If this ROM _is_ present (file size >0) the emulator sets
+            #   `jp $83DD` at RST 7 (patching TK80.ROM if present).
+            'BSMON.ROM':            # $F000-$FFFF
+                'https://gitlab.com/retroabandon/tk80-re/-/raw/main/rom/80BS-clone/etkroms3/bsmon.rom',
+            #   Only one of LV[12]BASIC.ROM is read based on the boot mode.
+            'LV1BASIC.ROM':         # $E000-$EFFF
+                'https://gitlab.com/retroabandon/tk80-re/-/raw/main/rom/80BS-clone/etkroms/L1CB.ROM',
+            'LV2BASIC.ROM':         # $D000-$EFFF
+                'https://gitlab.com/retroabandon/tk80-re/-/raw/main/rom/80BS-clone/etkroms3/LV2BASIC.ROM',
+            'FONT.ROM':             # chargen addr space
+                'https://gitlab.com/retroabandon/tk80-re/-/raw/main/rom/80BS-clone/etkroms3/font.rom',
+        },
+        'tk85': {
+            'TK85.ROM':             # $0000-$07FF fill $FF
+                'https://gitlab.com/retroabandon/tk80-re/-/raw/main/rom/TK85.bin',
+            'EXT.ROM': None,        # $0C00-$7BFF
+        },
+    }
 
 class VICE(Suite):
-    pass
+    suitedesc = 'CBM/Commdore 8-bit systems emulators'
 
 class OpenMSX(Suite):
-    pass
+    suitedesc = 'MSX and related systems emulators'
 
 class Linapple(Suite):
-    pass
+    suitedesc = 'Apple II emulator'
 
 class RunCPM(Suite):
     #   XXX This requires that RunCPM be in the path. We should check this
     #   and suggest `t8dev buildtoolset RunCPM` if it's not present.
+
+    suitedesc = 'RunCPM CP/M 2.x emulator'
+
+    @classmethod
+    def setargparser(cls, parser):
+        parser.add_argument('-a', '--autorun', action='store_true',
+            help='automatically run the .COM file given as first argument')
+        parser.add_argument('file', nargs='*',
+            help='file to copy to A: drive')
 
     def run(self):
         #   XXX We use `=X` to specify option X, because `-X' will be
         #   consumed by the t8dev options parser. We need to get that
         #   options parser to use subcommands that can have their own
         #   separate options.
-        autorun = False
-        if self.args and self.args[0] == '=a':
-            autorun = True
-            self.args = self.args[1:]
-
         self.emudir = path.build('emulator', self.suitename())
-        self.setup_emudir(autorun)
+        self.setup_emudir(self.args.autorun)
         with cwd(self.emudir):
             #   RunCPM clears the screen on start, as well as doing other
             #   less annoying bits termios terminal manipulation. Set the
@@ -237,7 +246,7 @@ class RunCPM(Suite):
             for drive in drives:  drive.mkdir(exist_ok=True, parents=True)
 
         #   Copy specified files to drive A.
-        for file in map(Path, self.args):
+        for file in map(Path, self.args.file):
            self.copycpmfile(file, A0)
 
         #   Copy standard CP/M commands to drive C, if they are present.
@@ -257,7 +266,7 @@ class RunCPM(Suite):
             #   Using XXX.COM to run a program in RunCPM often works, but
             #   not always; e.g. `TMON100.COM` can't seem to find that
             #   file, while `TMON100` works. So make sure we drop `.COM`.
-            commands = [Path(self.args[0]).stem.upper(), 'EXIT']
+            commands = [Path(self.args.file[0]).stem.upper(), 'EXIT']
             vprint(1, 'RunCPM', f'autorun: {commands}')
             subdata = compile_submit(commands)
             with open(self.emudir.joinpath(A0, '$$$.SUB'), 'wb') as f:
@@ -304,6 +313,8 @@ class RunCPM(Suite):
 
 
 ####################################################################
+
+SUITE_CLASSES = [ CSCP, Linapple, OpenMSX, RunCPM, VICE ]
 
 SUITES = dict([ (s.suitename(), s) for s in [
     CSCP, Linapple, OpenMSX, RunCPM, VICE]])