+++ psrc/t8dev/cli/t8dev/main.py
@@ -1,79 +1,68 @@
 ''' cli.t8dev argument parsing and top level '''
 
 from    argparse  import ArgumentParser
+from    pathlib  import Path
 from    site  import addsitedir
 import  os, sys
 
-from    t8dev.cli.t8dev.build  import *
-from    t8dev.cli.t8dev.emulator  import emulator
-from    t8dev.cli.t8dev.execute  import *
-from    t8dev.cli.t8dev.toolset  import buildtoolsets, buildtoolset
+from    t8dev  import path
 from    t8dev.cli.t8dev.util  import vprint
-from    t8dev.cli import exits
 import  t8dev.cli.t8dev.shared as shared
+
+from    t8dev.cli.t8dev.toolset     import setargs_toolset
+from    t8dev.cli.t8dev.asl         import setargs_asl
+from    t8dev.cli.t8dev.asx         import setargs_asx
+from    t8dev.cli.t8dev.pytest      import setargs_pytest
+from    t8dev.cli.t8dev.build       import setargs_a2dsk
+from    t8dev.cli.t8dev.emulator    import setargs_emulator
 
 def parseargs():
     ''' Parse arguments. If any of the arguments generate help messages,
         this will also exit with an appropriate code.
     '''
-    p = ArgumentParser(
-        description='Tool to build toolsets and code for 8-bit development.')
-    a = p.add_argument
-    a('-E', '--exclude', default=[], action='append',
-        help='for commands that do discovery, exclude these files/dirs'
-             ' (can be specified multiple times)')
+    ptop = ArgumentParser(description=
+        'Tool to build toolsets and code for 8-bit development.')
+    a = ptop.add_argument
     a('-P', '--project-dir',
         help='project directory; overrides T8_PROJDIR env var')
     a('-v', '--verbose', action='count', default=0,
         help='increase verbosity; may be used multiple times')
-    a('command', help="command; 'help' to list commands")
-    a('args', nargs='*',
-        help="arguments to command (preceed with '--' to use args with '-')")
 
-    args = p.parse_args()
-    if not COMMANDS.get(args.command):
-        exits.arg(f"t8dev: bad command '{args.command!r}';"
-            " use 'help' to list commands")
+    command_group = ptop.add_subparsers(dest='command',
+        # We do not use required=True; see below
+        title='Subcommands', metavar='', help='')
+    setargs_toolset(command_group)
+    setargs_asx(command_group)
+    setargs_asl(command_group)
+    setargs_pytest(command_group)
+    setargs_a2dsk(command_group)
+    setargs_emulator(command_group)
+
+    #   XXX Without a subcommand, this produces a poor error message:
+    #   't8dev: error: the following arguments are required:' (with nothing
+    #   after that). Ideally we should print the top-level help message as
+    #   produced by ptop.format_help(), or at least say we need a
+    #   subcommand and print a list of subcommands. It's not clear how to
+    #   do that without some messy catching of SystemExit from
+    #   ptop.parse_args().
+
+    args = ptop.parse_args()
+    if not args.command:
+        cmd = Path(sys.argv[0]).name
+        print( f'{cmd}: subcommand required; use -h for help', file=sys.stderr)
+        #   We don't print the help here because it's rather verbose.
+        exit(2)
     return args
 
-def help_commands(_):
-    print('Commands:')
-    for c in sorted(COMMANDS):
-        print('  ', c)
-
-COMMANDS = {
-    'help':     help_commands,
-
-    'asl':      asl,                # Assemble single program with ASL
-    'asl-testrig':  asl_testrig,    # Create and build source for a pytest
-                                    #   module that uses a `test_rig`.
-    'aslauto':  aslauto,            # Discover and build w/ASL all files used by
-                                    # unit tests (`object_files`/`test_rig`)
-    'aslt8dev': aslt8dev,           # Build assembly code included with t8dev.
-
-    'asx':      asx,                # ASXXXX assembler
-    'asxlink':  asxlink,            # ASXXXX linker
-
-    'a2dsk':    a2dsk,              # Apple II .dsk image that boots
-                                    #   and runs program
-
-    'bt':       buildtoolset,  'buildtoolset':  buildtoolset,
-    'bts':      buildtoolsets, 'buildtoolsets': buildtoolsets,
-
-    'pytest':   pytest,
-
-    'e':        emulator, 'emulator': emulator,
-}
-
 def main():
-    shared.ARGS = parseargs()
+    args = shared.ARGS = parseargs()
 
-    if shared.ARGS.project_dir:    # override environment
-        path.T8_PROJDIR = path.strict_resolve(shared.ARGS.project_dir)
+    if args.project_dir:    # override environment
+        path.T8_PROJDIR = path.strict_resolve(args.project_dir)
     if path.T8_PROJDIR is None:
         raise RuntimeError('T8_PROJDIR not set')
-    vprint(1, '========',
-        't8dev command={} args={}'.format(shared.ARGS.command, shared.ARGS.args))
+    os.environ['T8_PROJDIR'] = str(path.T8_PROJDIR)
+    vprint(1, f'━━━━━━━━ t8dev', args.command)
     vprint(1, 'projdir', str(path.proj()))
 
     #   Code common to several .pt files (including entire test suites) may
@@ -86,5 +75,4 @@ def main():
     #   to sit down and work out how we really want to deal with it.
     #
     addsitedir(str(path.proj()))
-    cmdf = COMMANDS.get(shared.ARGS.command)
-    exit(cmdf(shared.ARGS.args))
+    exit(args.func(args))